Right thumb layer key now goes to numbers layer by default
This breaks the symmetry of the two halves of the Ergodox, but who am
I kidding, the two halves are already different.

With this change, the layer thumb key on the left hand accesses the
symbols layer on one tap and the numbers layer on the second.  The
right hand layer thumb key accesses the numbers layer on one tap and
the media layer on the second.

Now I'm inspired to create a new layer for the second tap on the left
thumb key.  God help me.