Make layer locking harder to activate on accident
Before, locking a layer on was done just like with TT, one tap locked
a layer on.  But I found that I was doing that too often on accident.

Now you have to tap with the shift key held down.  The nice thing
about this is you can jump around layers while the firmware thinks you
are in layer lock mode.